pyproject.toml
[project]
name = "nonebot-plugin-gkmscalculator"
version = "0.2.0"
description = "学园偶像大师算分：产出评级、训练属性、强化月算分与角色别称管理"
authors = [{ name = "病毒", email = "769185426@qq.com" }]
dependencies = [
    "nonebot2>=2.3.1",
    "nonebot-adapter-onebot>=2.4.0",
    "pydantic>=2.0.0",
    "httpx>=0.27.0",
    "numpy>=1.24.0",
    "opencv-python-headless>=4.8.0",
]
requires-python = ">=3.9"
readme = "README.md"
license = { text = "MIT" }
keywords = ["nonebot", "nonebot2", "plugin", "gakumas", "学马"]

[project.urls]
Homepage = "https://pypi.org/project/nonebot-plugin-gkmscalculator/"

[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"

[tool.pdm]
distribution = true

[tool.pdm.build]
includes = ["src"]
# 发布包不包含测试与 CI 配置
excludes = ["tests", "tests/**", ".github", ".github/**"]

[tool.pdm.dev-dependencies]
dev = [
    "black>=24.4.2",
    "isort>=5.13.2",
    "ruff>=0.4.6",
    "pytest>=8.0.0",
]

[tool.pdm.scripts]
format = { composite = ["isort .", "black .", "ruff check ."] }
test = "pytest tests -q"

[tool.black]
line-length = 90
target-version = ["py39", "py310", "py311", "py312"]
include = '\.pyi?$'
extend-exclude = '''
'''

[tool.isort]
profile = "black"
line_length = 90
length_sort = true
skip_gitignore = true
force_sort_within_sections = true
extra_standard_library = ["typing_extensions"]

[tool.ruff]
line-length = 90
target-version = "py39"

[tool.ruff.lint]
select = ["E", "W", "F", "UP", "C", "T", "PYI", "PT", "Q"]
ignore = ["C901", "T201", "E731", "E402"]

[tool.pyright]
pythonVersion = "3.9"
pythonPlatform = "All"
typeCheckingMode = "basic"

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]

[tool.nonebot]
adapters = [
    { name = "OneBot V11", module_name = "nonebot.adapters.onebot.v11", project_link = "https://github.com/nonebot/adapter-onebot" },
]
plugin_dirs = ["src"]
